필터 수정(확인요망), SRP법칙 적용, TODO 설정,으로 HPObject 추가,

+++ Server/ChessServer/ChessServer/Packet.h
@@ -126,8 +126,8 @@ namespace chess::packet
 
     struct SC_PACKET_ATTACK : PacketHeader
 	{
-        uint32_t    _attacker_id;
-        uint32_t    _target_id;
+        int64_t     _attacker_id;
+        int64_t     _target_id;
         int32_t     _damage;
         int32_t     _target_current_hp;
     };
